requirements.txt
python-snappy==0.7.3
seri @ git+https://github.com/aroxby/seri.git@e0458ff1799ec8975275906b86c98b75cb2c4a2d
